backend/requirements.txt
# Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0  # Required for EmailStr validation

# WebSocket support
websockets>=13.0

# Redis caching
redis>=5.0.0
hiredis>=2.3.0

# Code Analysis
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0

# AI/ML
openai>=1.54.0
pinecone>=5.0.0
tiktoken>=0.8.0

# Utils
gitpython>=3.1.0
aiofiles>=24.0.0

# Supabase (postgrest-py handled as dependency)
supabase>=2.0.0
pyjwt>=2.8.0  # JWT token verification for Supabase Auth

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# Observability
sentry-sdk[fastapi]>=2.0.0

# Search V2 - Hybrid search
rank-bm25>=0.2.2
cohere>=5.0.0

# Search V3 - Code-optimized embeddings
voyageai>=0.3.0
